Parse the PVH entry point from the ELF header
The PVH entry point is parsed from the note segment in the ELF
header and returned. Debugging statements are enabled using logger
and warn!() macro in several modules. Note that it is needed to
modify the Cargo.toml file of the specific module and then add the
use extern crate to the top module. See diff for examples.
Next step is to correctly set up the start_info struct, using the
example for the bootparams that is the current method

Likely problems are the paging setup and the value of the task
register as required by the ABI:
https://xenbits.xen.org/docs/4.8-testing/misc/hvmlite.html