Start populating the hvm_start_info struct.
Wrote a basic (must replace) mechanism to determine when PVH
boot is being used and fill out the hvm_start_info struct
instead of the typical boot_params.
Created the start_info and related structs and started to
populate the required values. Next to do is to populate the
E820 maps into the corresponding fields of the structs, and
set up the registers correctly.